+++ packages/firebase-dart/lib/src/firestore.dart
@@ -734,12 +734,20 @@ class QuerySnapshot
   /// Non-null list of the documents that changed since the last snapshot.
   /// If this is the first snapshot, all documents will be in the list as
   /// added changes.
-  List<DocumentChange> get docChanges =>
-      jsObject.docChanges.map(DocumentChange.getInstance).toList();
+  List<DocumentChange> get docChanges => jsObject.docChanges
+      // explicitly casting the returned list due to
+      // https://github.com/dart-lang/sdk/issues/33537
+      .cast<firestore_interop.DocumentChangeJsImpl>()
+      .map(DocumentChange.getInstance)
+      .toList();
 
   /// Non-null list of all the documents.
-  List<DocumentSnapshot> get docs =>
-      jsObject.docs.map(DocumentSnapshot.getInstance).toList();
+  List<DocumentSnapshot> get docs => jsObject.docs
+      // explicitly casting the returned list due to
+      // https://github.com/dart-lang/sdk/issues/33537
+      .cast<firestore_interop.DocumentSnapshotJsImpl>()
+      .map(DocumentSnapshot.getInstance)
+      .toList();
 
   /// [:true:] if there are no documents.
   bool get empty => jsObject.empty;