+++ packages/firebase-dart/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Updated tested JS API version to `4.11.0`.
 
+* Require at least Dart SDK `2.0.0-dev.36`.
+
 ### Firestore
 
 * Moved `setLogLevel` to a top-level method from `Firestore`.