+++ packages/firebase-dart/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 4.5.0
 
-* Updated tested JS API version to `4.11.0`.
+* Updated tested JS API version to `4.13.0`.
 
 * Require at least Dart SDK `2.0.0-dev.36`.